[emsymbolizer] Move functions to tools/symbolizer.py (NFC)
This way the functions can be possibly reused from files from `tools/`.
